Cleaning of datastore's plugin.py
