+++ ckanext/datastore/tests/test_unit.py
@@ -9,6 +9,7 @@ from ckan.common import config
 import ckan.tests.legacy as tests
 import ckanext.datastore.db as db
 import ckanext.datastore.plugin as plugin
+import ckanext.datastore.backend as backend
 
 
 class TestTypeGetters(unittest.TestCase):
@@ -53,7 +54,7 @@ class TestLegacyModeSetting():
             'ckan.datastore.write_url': config['ckan.datastore.write_url'],
         }
 
-        assert plugin._is_legacy_mode(test_config)
+        assert backend._is_legacy_mode(test_config)
 
     @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
     def test_legacy_mode_set_if_no_read_url_and_pg_8(self, pgv):
@@ -64,7 +65,7 @@ class TestLegacyModeSetting():
             'ckan.datastore.write_url': config['ckan.datastore.write_url'],
         }
 
-        assert plugin._is_legacy_mode(test_config)
+        assert backend._is_legacy_mode(test_config)
 
     @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
     def test_legacy_mode_set_if_read_url_and_pg_8(self, pgv):
@@ -76,7 +77,7 @@ class TestLegacyModeSetting():
             'ckan.datastore.read_url': 'some_test_read_url',
         }
 
-        assert plugin._is_legacy_mode(test_config)
+        assert backend._is_legacy_mode(test_config)
 
     @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
     def test_legacy_mode_not_set_if_read_url_and_pg_9(self, pgv):
@@ -88,4 +89,4 @@ class TestLegacyModeSetting():
             'ckan.datastore.read_url': 'some_test_read_url',
         }
 
-        assert not plugin._is_legacy_mode(test_config)
+        assert not backend._is_legacy_mode(test_config)
